Update from releng/14.1, git revision 1a207e5cdf99.

+++ usr.bin/patch/pch.h
@@ -24,7 +24,6 @@
  * behaviour
  *
  * $OpenBSD: pch.h,v 1.9 2003/10/31 20:20:45 millert Exp $
- * $FreeBSD: 53ed71780e2674aedb915ef4be37e3fb74196d9b $
  */
 
 #define	OLD_FILE	0
@@ -46,7 +45,7 @@ bool		there_is_another_patch(void);
 bool		another_hunk(void);
 bool		pch_swap(void);
 char		*pfetch(LINENUM);
-unsigned short	pch_line_len(LINENUM);
+size_t		pch_line_len(LINENUM);
 LINENUM		pch_first(void);
 LINENUM		pch_ptrn_lines(void);
 LINENUM		pch_newfirst(void);
